updated code for guestbook to include assets
