+++ develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-service/bnd.bnd
@@ -3,6 +3,6 @@ Bundle-SymbolicName: com.liferay.docs.guestbook.service
 Bundle-Version: 1.0.0
 Liferay-Require-SchemaVersion: 1.0.0
 Liferay-Service: true
-Export-Package: 
-    com.liferay.docs.guestbook.asset,\
-	com.liferay.docs.guestbook.search
+Export-Package:
+	com.liferay.docs.guestbook.service.permission, \
+	com.liferay.docs.guestbook.search,\
