+++ develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-service/src/main/resources/service.properties
@@ -13,5 +13,5 @@
 ##
 
     build.namespace=GB
-    build.number=22
-    build.date=1547656172326
+    build.number=38
+    build.date=1550249999256
