+++ develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-web/bnd.bnd
@@ -1,4 +1,4 @@
 Bundle-Name: guestbook-web
 Bundle-SymbolicName: com.liferay.docs.guestbook.portlet
 Bundle-Version: 1.0.0
-Export-Package: com.liferay.docs.guestbook.portlet.constants
+Export-Package:
