+++ develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-web/src/main/java/com/liferay/docs/guestbook/web/internal/security/permission/resource/GuestbookPermission.java
