+++ develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-web/src/main/resources/META-INF/resources/asset/full_content.jsp
@@ -1,4 +1,19 @@
-<%@include file="../../init.jsp"%>
+ <%@include file="../../init.jsp"%>
+
+        <%
+        Guestbook guestbook = (Guestbook)request.getAttribute("gb_guestbook");
+
+        guestbook = guestbook.toEscapedModel();
+        %>
+
+        <dl>
+                <dt>Name</dt>
+                <dd><%= guestbook.getName() %></dd>
+        </dl>
+        
+                request.setAttribute("gb_guestbook", _guestbook);
+        
+          <%@include file="../../init.jsp"%>
 
         <%
         Entry entry = (Entry)request.getAttribute("gb_entry");
