--- develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-web/src/main/resources/META-INF/resources/asset/guestbook/full_content.jsp
@@ -1,12 +1,0 @@
-<%@include file="../../init.jsp"%>
-
-        <%
-        Guestbook guestbook = (Guestbook)request.getAttribute("gb_guestbook");
-
-        guestbook = guestbook.toEscapedModel();
-        %>
-
-        <dl>
-                <dt>Name</dt>
-                <dd><%= guestbook.getName() %></dd>
-        </dl>
