+++ develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-web/src/main/resources/META-INF/resources/guestbookadminportlet/edit_guestbook.jsp
@@ -26,33 +26,38 @@
                 <aui:fieldset>
                      <aui:input name="name" />
                 </aui:fieldset>
-
+				<aui:fieldset>
+		<liferay-asset:asset-categories-selector
+			className="<%= Guestbook.class.getName() %>"
+			classPK="<%= guestbookId %>" />
+		<liferay-asset:asset-tags-selector
+			className="<%= Guestbook.class.getName() %>"
+			classPK="<%= guestbookId %>" />
+	</aui:fieldset>
                 <aui:button-row>
                      <aui:button type="submit" />
                      <aui:button onClick="<%= viewURL %>" type="cancel"  />
                 </aui:button-row>
                 
-        <liferay-ui:asset-categories-error />
-        <liferay-ui:asset-tags-error />
-        
-          <liferay-ui:panel defaultState="closed" extended="<%= false %>"
-          id="guestbookCategorizationPanel" persistState="<%= true %>"
-          title="categorization">
-			<aui:fieldset>
-				
-				<liferay-asset:asset-categories-selector className="<%= Guestbook.class.getName() %>" classPK="<%= guestbookId %>" />
-				
-				<liferay-asset:asset-tags-selector className="<%= Guestbook.class.getName() %>" classPK="<%= guestbookId %>" />
-            </aui:fieldset>
-        </liferay-ui:panel>
-        
-         <liferay-ui:panel defaultState="closed" extended="<%= false %>"
-          id="guestbookAssetLinksPanel" persistState="<%= true %>"
-          title="related-assets">
-          <aui:fieldset>
-            <liferay-ui:input-asset-links
-              className="<%= Guestbook.class.getName() %>"
-              classPK="<%= guestbookId %>" />
-          </aui:fieldset>
-        </liferay-ui:panel>
-        </aui:form>
+                 <liferay-ui:asset-categories-error />
+	<liferay-ui:asset-tags-error />
+	<liferay-ui:panel defaultState="closed" extended="<%= false %>"
+		id="guestbookCategorizationPanel" persistState="<%= true %>"
+		title="categorization">
+		 <aui:fieldset>
+           <liferay-asset:asset-categories-selector className="<%= Guestbook.class.getName() %>" classPK="<%= guestbookId %>" />
+           <liferay-asset:asset-tags-selector className="<%= Guestbook.class.getName() %>" classPK="<%= guestbookId %>" />
+         </aui:fieldset>
+	</liferay-ui:panel>
+
+	<liferay-ui:panel defaultState="closed" extended="<%= false %>"
+		id="guestbookAssetLinksPanel" persistState="<%= true %>"
+		title="related-assets">
+		<aui:fieldset collapsed="<%= true %>" collapsible="<%= true %>"
+			label="related-assets">
+			<liferay-asset:input-asset-links
+				className="<%= Guestbook.class.getName() %>"
+				classPK="<%= guestbookId %>" />
+		</aui:fieldset>
+	</liferay-ui:panel>
+</aui:form>
