+++ develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-web/src/main/resources/META-INF/resources/guestbookwebportlet/entry_actions.jsp
@@ -10,7 +10,7 @@
 
         <liferay-ui:icon-menu>
         
-          <portlet:renderURL var="viewEntryURL">
+        <portlet:renderURL var="viewEntryURL">
       <portlet:param name="entryId"
         value="<%= String.valueOf(entry.getEntryId()) %>" />
       <portlet:param name="mvcPath"
@@ -18,16 +18,7 @@
     </portlet:renderURL>
 
     <liferay-ui:icon message="View" url="<%= viewEntryURL.toString() %>" />
-
-            <portlet:renderURL var="viewEntryURL">
-                <portlet:param name="entryId" value="<%= String.valueOf(entry.getEntryId()) %>" />
-                <portlet:param name="mvcPath" value="/guestbookwebportlet/view_entry.jsp" />
-            </portlet:renderURL>
-
-            <liferay-ui:icon
-                message="View"
-                url="<%= viewEntryURL.toString() %>"
-            />
+        
 
             <c:if
                 test="<%= GuestbookEntryPermission.contains(permissionChecker, entry.getEntryId(), ActionKeys.UPDATE) %>">
