+++ develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-web/src/main/resources/META-INF/resources/init.jsp
@@ -6,17 +6,24 @@
 taglib uri="http://liferay.com/tld/portlet" prefix="liferay-portlet" %><%@
 taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
 taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
-<%@ taglib uri="http://liferay.com/tld/asset" prefix="liferay-asset" %>
-<%@ taglib uri="http://liferay.com/tld/aui" prefix="aui" %>
 
+ <%@ taglib uri="http://liferay.com/tld/asset" prefix="liferay-asset" %> 
+<%@ page import="java.util.Map" %> <%@ page import="java.util.HashMap" %>
 
+<%@ page import="com.liferay.asset.kernel.service.AssetEntryLocalServiceUtil" %>
+<%@ page import="com.liferay.asset.kernel.service.AssetTagLocalServiceUtil" %>
+
+<%@ page import="com.liferay.asset.kernel.model.AssetEntry" %>
+<%@ page import="com.liferay.asset.kernel.model.AssetTag" %>
+
+<%@ page import="com.liferay.portal.kernel.util.ListUtil" %>
 <%@ taglib uri="http://liferay.com/tld/frontend" prefix="liferay-frontend" %>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 <%@ taglib uri="http://liferay.com/tld/security" prefix="liferay-security" %>
 <%@ page import="java.util.List" %>
 <%@ page import="com.liferay.portal.kernel.util.ParamUtil" %>
 <%@ page import="com.liferay.portal.kernel.util.HtmlUtil" %>
-<%@ page import="com.liferay.portal.kernel.util.StringPool" %>
+<%@ page import="com.liferay.petra.string.StringPool" %>
 <%@ page import="com.liferay.portal.kernel.model.PersistedModel" %>
 <%@ page import="com.liferay.portal.kernel.dao.search.SearchEntry" %>
 <%@ page import="com.liferay.portal.kernel.dao.search.ResultRow" %>
@@ -45,15 +52,9 @@ taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 <%@ page import="com.liferay.portal.kernel.util.GetterUtil" %>
 <%@ page import="com.liferay.portal.kernel.util.Validator" %>
 <%@ page import="com.liferay.portal.kernel.util.PortalUtil" %>
-<%@ page import="java.util.Map" %>
-<%@ page import="java.util.HashMap" %>
-
-<%@ page import="com.liferay.asset.kernel.service.AssetEntryLocalServiceUtil" %>
-<%@ page import="com.liferay.asset.kernel.service.AssetTagLocalServiceUtil" %>
-
-<%@ page import="com.liferay.asset.kernel.model.AssetEntry" %>
-<%@ page import="com.liferay.asset.kernel.model.AssetTag" %>
-<%@ page import="com.liferay.portal.kernel.util.ListUtil" %>
+<%@page import="com.liferay.portal.kernel.comment.Discussion" %>
+<%@page import="com.liferay.portal.kernel.comment.CommentManagerUtil" %>
+<%@page import="com.liferay.portal.kernel.service.ServiceContextFunction" %>
 
 <%@ page import="java.util.ArrayList" %>
 <%@ page import="java.util.Map" %>
