+++ develop/tutorials/code/com-liferay-docs-guestbook/modules/guestbook/guestbook-web/src/main/resources/content/Language.properties
